refactor(packages/awareness): consolidate payload guards in types/guards
The WebSocket and BroadcastChannel adapters carried byte-identical copies of
isRecord / isCursorPosition / isSelectionRange / isPointerPosition. Lift
them into src/types/guards.ts so the wire shape lives in one place and
future fields cannot drift between adapters. The adapter-specific
isPresenceUser and update-payload guards stay local since they validate
different field sets.
